+++ core-foundation-sys/src/lib.rs
@@ -12,3 +12,4 @@ pub mod number;
 pub mod runloop;
 pub mod set;
 pub mod string;
+pub mod url;
